+++ NEWS.md
@@ -1,3 +1,8 @@
+# CHANGES TO prediction 0.1.5
+
+* Added `prediction.svm()` method for objects of class `"svm"` from **e1071**. (#1)
+* Fixed a bug in `prediction.polr()` when attempting to pass a `type` argument, which is always ignored. A warning is now issued when attempting to override this.
+
 # CHANGES TO prediction 0.1.4
 
 * Added `mean_or_mode()` and `median_or_mode()` functions, which provide a simple way to aggregate a variable of factor or numeric type. (#3)